Refactor social share widget into re-usable component.
